Course - Section 2 - Task Management Application (REST API)
This part is to now the bases of Nest.js without using a database, only with arrays.

key points:
What are the available properties for the @Module() decorator?
Import, export, controllers and providers.

What is the first argument of the @Controller() decorator?
It's a string, describing the path/route that the conroller handles

What are Data Transfer Objects used for?
DTOs define the shape of data of an incoming request, and allows us to re-use the definition throughout the application.

How would you retrieve the entire body parameters of a request?
By prefixing the handler parameter with the @Body decorator, without providing an argument.

How would you define a handler that can retrieve the "id" parameter from a URL?
I'll add '/:id' to the handler's endpoint string definition.
Then I can retrieve it using '@Param("id")'.

What is the @Injectable() decorator used for?
It's used to define that a certain class should have a sheared instance across the module.
The instance can then be injected using Dependency injection, and all injectors will have access to the same instance and its state.