Section 3: Validation and Error Handling
Pipes are used to validate the data, which is inserted in a request before executing any process to avoid errors in the function.

It's important to add the below line in main.ts
app.useGlobalPipes(new ValidationPipe());

We can use the DTO's to implement pipes that's why we need to install next libraries:
yarn add class-validator class-transform

We can use multiple validation for value
Exemple:

@IsOptional
@IsEnum(TaskStatus)
status? : TaskStatus